crews.go

+++ crews.go
@@ -35,7 +35,7 @@ func (c *Crews) ReportHttp(s string) *Crews {
 func (c *Crews) PrimaryTest(s string) bool {
 	count := c.primary.Test([]byte(s))
 	adjustedCount := uint((float64(count) - c.p * float64(c.primary.count))/(c.q-c.p))
-	return  adjustedCount >= uint(( c.primary.count/c.numWebsites)*c.primaryThresholdModifier)
+	return  adjustedCount >= uint(float64(( c.primary.count/c.numWebsites))*c.primaryThresholdModifier)
 }
 
 func (c *Crews) SecondaryTest(s string) bool {
